+++ classes.py
@@ -1,8 +1,20 @@
 """module with classes"""
 from datetime import datetime
 import sqlite3
-from functions import show_message, get_random_string
+from tkinter import messagebox
+import random
+import string
 
+
+def show_message(title, message):
+    """Pokazuje wiadomosci oraz bledy"""
+    messagebox.showerror(title, message)
+
+
+def get_random_string(self):
+    """generuje losowy oraz unikatowy id"""
+    letters = string.ascii_lowercase
+    return ''.join(random.choice(letters) for i in range(8))
 
 class Event:
     """Przodek klas związanych z wydarzeniem"""
