+++ main.py
@@ -1,10 +1,19 @@
-from functions import load_from_json
-
+import os
+from classes import Event, Participant, EventCreator
+from functions import get_list_from_json, make_test_jsons
 
 
 if __name__ == '__main__':
 
-    event_creator_list = load_from_json(EventCreator, 'jsons/event_creator_list.json')
-    event_list = load_from_json(Event, 'jsons/event_list.json')
-    participant_list = load_from_json(Participant, 'jsons/participant.json')
+    if not os.path.exists('jsons/*.json'):
+        make_test_jsons()
+
+    event_creator_list = get_list_from_json(EventCreator, 'jsons/event_creator_list.json')
+    event_list = get_list_from_json(Event, 'jsons/event_list.json')
+    participant_list = get_list_from_json(Participant, 'jsons/participant.json')
+
+    print(f"{event_list = }")
+    print(f"{event_creator_list = }")
+    print(f"{participant_list = }")
+
